# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: Grid Up — Pano Beyni Merkez API
  # v1.1.0 (18 Eylul 2026): GET /fleet/health eklendi. Yalnizca EKLEME, geri uyumlu.
  #   Gerekce: contracts/changes/2026-09-14-fleet-health-bulk.md
  # v1.2.0 (18 Eylul 2026): onay/raf govdesinden `by` KALKTI, securitySchemes eklendi (F-19).
  #   GERI UYUMLU DEGIL: onaylayanin adi artik govdeden degil Authorization basligindan
  #   turer. Eski istemcinin govdeye yazdigi `by` sessizce YOK SAYILIR (reddedilmez).
  #   Gerekce: contracts/changes/2026-09-18-kimlik-dogrulama.md
  # v1.3.0 (18 Eylul 2026): varlik kutugu (F-21). AssetRegistry semasi, PanelDetail.asset,
  #   PanelSummary'ye uc alan (abone_sayisi, kritiklik, sonraki_bakim_at) ve GET/POST
  #   /fleet/assets. Yalnizca EKLEME, geri uyumlu: yeni alanlarin hicbiri required degil ve
  #   kunye ice aktarilmamissa null doner.
  #   Gerekce: contracts/changes/2026-09-18-varlik-kutugu.md
  # v1.4.0 (18 Eylul 2026): ust sebeke kesintisi bagintisi (F-22). OutageEvent semasi,
  #   GET /outages + GET /outages/{outage_id}, Alarm'a outage_id. Yalnizca EKLEME, geri
  #   uyumlu: outage_id required degil ve baginti kurulmayan alarmda bulunmaz.
  #   Gerekce: contracts/changes/2026-09-18-kesinti-bagintisi.md
  version: 1.4.0
  description: |
    On-prem izleme platformunun HTTP arayuzu. Tamami sirket ici aginda calisir,
    public cloud bagimliligi yoktur (PLAN.md GK4).
servers:
  - url: http://localhost:8000
    description: Gelistirme (dogrudan backend)
  - url: /
...
paths:
...
  /api/v1/alarms:
    get:
      tags: [alarms]
      summary: Alarm konsolu listesi
      parameters:
        - { name: state, in: query, schema: { type: string, default: "active,acked" }, description: "active,acked,shelved,cleared" }
        - { name: prio, in: query, schema: { type: string }, description: "P1,P2,P3,INFO,SYS" }
        - { name: pano_id, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 100 } }
      responses:
        "200":
          description: Alarmlar (en yeni once)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Alarm" }
  /api/v1/alarms/{alarm_id}/ack:
    post:
      tags: [alarms]
      summary: Alarm onayi (ISA-18.2)
      parameters:
        - $ref: "#/components/parameters/AlarmId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                note: { type: string }
                channel: { type: string, enum: [ui, sms, scada], default: ui }
      security:
        - operatorToken: []
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "Zaten onaylanmis veya temizlenmis"
  /api/v1/alarms/{alarm_id}/shelve:
    post:
      tags: [alarms]
      summary: Alarmi süreli ve gerekceli rafa al (P1 rafa alinamaz)
      parameters:
        - $ref: "#/components/parameters/AlarmId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [minutes, reason]
              properties:
                minutes: { type: integer, minimum: 1, maximum: 480 }
                reason: { type: string, minLength: 3 }
      security:
        - operatorToken: []
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: >
            P1 alarm rafa alinamaz (alarm-codes.yaml suppressible=false),
            VEYA belirtecin rolu yetmiyor ('operator' gerekir).
...
  /api/v1/outages:
    get:
      tags: [events]
      summary: Ust sebeke kesintileri (es zamanli susan panolar tek olayda)
      description: >
        Ayni FIDERDEKI panolarin es zamanli susmasi N ayri ariza degil TEK bir kesinti
        olayidir (F-22). Baginti `panels.fider_id` uzerinden kurulur (F-21 kunyesi);
        fider bilgisi olmayan pano GRUPLANMAZ. Esikler alarm-codes.yaml
        `outage_min_panels` / `outage_window_min`.
      parameters:
        - name: state
          in: query
          description: "acik = suren kesintiler; hepsi = kapananlar dahil"
          schema: { type: string, enum: [acik, hepsi], default: acik }
      responses:
        "200":
          description: Kesinti olaylari, en yenisi once
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/OutageEvent" }
  /api/v1/outages/{outage_id}:
    get:
      tags: [events]
      summary: Tek kesinti olayi
      parameters:
        - { name: outage_id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Kesinti olayi
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OutageEvent" }
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
...
    AlarmId:
      name: alarm_id
      in: path
      required: true
      schema: { type: string }
  responses:
    Ok:
      description: Basarili
      content:
        application/json:
          schema:
            type: object
            properties:
              ok: { type: boolean }
    NotFound:
      description: Bulunamadi
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: >
        Belirtec yok, bicimi desteklenmiyor veya gecersiz. `WWW-Authenticate: Bearer`
        basligi doner. NOT: operator tablosu bos birakilirsa kimlik dogrulama KAPALIDIR
        ve bu yanit hic uretilmez; durum GET /health `auth.enabled` alanindan okunur.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: "Kimlik dogru ama rol yetmiyor (izleyici < operator < muhendis)"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  securitySchemes:
    operatorToken:
      type: http
      scheme: bearer
      description: >
        On-prem operator belirteci (F-19). Kurumsal SSO/OIDC DEGILDIR: belirtecler
        yapilandirmada duran paylasilan sirlardir; parola, oturum suresi, yenileme ve
        iptal listesi yoktur. GK4 public cloud kimlik saglayicisini yasakladigi icin
        bulut IdP kullanilamaz. Yalnizca YAZMA uclarini (onay/raf) korur; okuma uclari
        ve WebSocket akisi belirtec istemez. Ayrinti: docs/15-guvenlik-kvkk.md §5.
  schemas:
    Error:
      type: object
      required: [detail]
      properties:
        detail: { type: string }
    OutageEvent:
      type: object
      description: >
        Ust sebeke kesintisi (F-22): ayni fiderde es zamanli susan N panonun TEK olayi.
        Bu bir ALARM DEGILDIR — alarmlari aciklayan bir olaydir; alt alarmlar
        (ALM-LASTGASP, ALM-COMMS-LOST) uretilmeye devam eder ve BASTIRILMAZ, yalnizca
        `outage_id` ile buna baglanir.
      required: [outage_id, fider_id, started_at, detected_at, panolar, state]
      properties:
        outage_id: { type: string, description: "or. OUT-F-EFELER-03-20260918T0900" }
        fider_id: { type: string, description: "kesintinin fideri (F-21 kunyesinden)" }
        started_at:
          type: string
          format: date-time
          description: >
            Panolarin sustugu an (kumedeki EN GEC son veri zamani). Bu, enerjinin
            kesildigi anin OLCULEBILEN en iyi yaklasimidir; kesme aninin kendisi
            gozlemlenmiyor.
        detected_at:
          type: string
          format: date-time
          description: "Merkezin bagintiyi kurdugu an (haberlesme zaman asimi kadar sonra)"
        ended_at:
          type: [string, "null"]
          format: date-time
          description: >
            Haberlesmenin GERI DONDUGU an. DIKKAT: bu enerjinin geri geldigi an DEGILDIR
            ve histerezislidir; restorasyon ani bu depoda olculmuyor (F-23).
        state: { type: string, enum: [acik, kapandi] }
        panolar:
          type: array
          description: "Olaya bagli panolar"
          items:
            type: object
            required: [pano_id, last_rx]
            properties:
              pano_id: { type: string }
              name: { type: string }
              last_rx: { type: string, format: date-time }
              abone_sayisi: { type: [integer, "null"], description: "kunyeden; null = bilinmiyor" }
        abone_toplami:
          type: [integer, "null"]
          description: >
            Etkilenen abone sayisinin TOPLAMI (EPDK Madde 8/2). Yalnizca kunyesi olan
            panolar toplanir ve `abone_eksik` kac panonun sayilamadigini soyler; hicbir
            panonun kunyesi yoksa null doner — SIFIR YAZILMAZ.
        abone_eksik:
          type: integer
          description: "abone sayisi bilinmediği icin toplama girmeyen pano sayisi"
...
    AlarmReason:
      type: object
      description: "Arayuzdeki 'Neden?' bolumu — her alarm aciklanabilir olmalidir (rapor 6.5 L4)"
      properties:
        signals:
          type: array
          items:
            type: object
            properties:
              tag: { type: string }
              value: { type: number }
              threshold: { type: number }
              unit: { type: string }
        layer: { type: string, enum: ["L-1", L0, L1, L2, L3] }
        basis: { type: string, description: "standart atfi, or. 'IEC 61439-1 Tablo 6'" }
        point: { type: [string, "null"] }
    Alarm:
      type: object
      required: [id, pano_id, code, prio, state, raised_at]
      properties:
        id: { type: string }
        event_id: { type: [string, "null"], description: "ayni kok nedenden gelen alarmlar bu id altinda gruplanir (PANO ICI)" }
        outage_id:
          type: [string, "null"]
          description: >
            Alarm bir ust sebeke kesintisine baglandiysa o kesintinin kimligi (F-22).
            event_id PANO ICI gruplamadir; bu alan PANOLAR ARASI bagintidir. Baginti
            alarmi BASTIRMAZ, yalnizca baglar.

        pano_id: { type: string }
        code: { type: string }
        text: { type: string, description: "Turkce alarm metni" }
        prio: { type: string, enum: [P1, P2, P3, INFO, SYS] }
        state: { type: string, enum: [active, acked, shelved, cleared] }
        raised_at: { type: string, format: date-time }
        cleared_at: { type: [string, "null"], format: date-time }
        acked_at: { type: [string, "null"], format: date-time }
        acked_by: { type: [string, "null"] }
        shelved_until: { type: [string, "null"], format: date-time }
        escalation_level: { type: integer, description: "0 = henuz eskalasyon yok" }
        notified: { type: array, items: { type: string, enum: [sms, whatsapp, call, scada, relay] } }
        reason: { $ref: "#/components/schemas/AlarmReason" }
        advice: { type: string, description: "'Ne yapmali?' — hypotheses[].advice" }
        ttl_h: { type: [number, "null"], description: "'Ne kadar acil?'" }
...